refactor: move type "RunningStatus" from "playback" to "lib"
