+++ tui/src/ui/mod.rs
@@ -37,11 +37,11 @@ use sysinfo::System;
 use termusiclib::player::music_player_client::MusicPlayerClient;
 use termusiclib::player::playlist_helpers::PlaylistRemoveTrackType;
 use termusiclib::player::PlayerProgress;
+use termusiclib::player::RunningStatus;
 use termusiclib::player::StreamUpdates;
 use termusiclib::player::UpdateEvents;
 use termusiclib::player::UpdatePlaylistEvents;
 pub use termusiclib::types::*;
-use termusicplayback::RunningStatus;
 use tokio::sync::mpsc::{self, UnboundedReceiver};
 use tokio_stream::Stream;
 use tokio_stream::StreamExt;
